requirements.txt
aiofiles==24.1.0
aistudio-sdk==0.3.6
annotated-types==0.7.0
anyio==4.10.0
appdirs==1.4.4
arxiv==2.2.0
async-timeout==3.0.1
asyncio==3.4.3
attrs==25.3.0
Authlib==1.6.3
backoff==2.2.1
bce-python-sdk==0.9.45
beautifulsoup4==4.13.4
blinker==1.9.0
boto3==1.40.27
botocore==1.40.27
Brotli==1.1.0
bubus==1.5.6
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
cfgv==3.4.0
chardet==3.0.4
charset-normalizer==3.4.2
click==8.2.1
cobble==0.1.4
coloredlogs==15.0.1
colorlog==6.9.0
cryptography==45.0.7
cssselect==1.3.0
cssselect2==0.8.0
Cython==3.1.3
decorator==5.2.1
defusedxml==0.7.1
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
docopt==0.6.2
docx==0.2.4
easyocr==1.7.2
EbookLib==0.18
einops==0.8.1
et_xmlfile==2.0.0
eutils==0.6.0
faiss-cpu==1.11.0.post1
fastapi==0.116.1
feedparser==6.0.12
ffmpeg==1.4
ffmpy==0.6.1
filelock==3.19.1
filetype==1.2.0
Flask==2.3.3
Flask-Cors==4.0.0
fonttools==4.59.2
frozendict==2.4.6
fsspec==2025.9.0
ftfy==6.3.1
future==1.0.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-genai==1.27.0
googleapis-common-protos==1.70.0
gradio==5.43.1
gradio_client==1.12.1
greenlet==3.2.4
groovy==0.1.2
groq==0.31.1
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
h2==4.3.0
habanero==2.3.0
hf-xet==1.1.9
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.30.2
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
humanfriendly==10.0
hyperframe==6.1.0
identify==2.6.14
idna==3.10
imageio==2.37.0
imagesize==1.4.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==0.3.22
langchain-anthropic==0.3.3
langchain-aws==0.2.19
langchain-core==0.3.49
langchain-deepseek==0.1.3
langchain-google-genai==2.1.2
langchain-ollama==0.3.0
langchain-openai==0.3.11
langchain-text-splitters==0.3.7
langsmith==0.3.45
lazy_loader==0.4
Levenshtein==0.27.1
libmagic==1.0
lxml==6.0.1
lxml_html_clean==0.4.2
mammoth==1.10.0
Markdown==3.9
markdown-it-py==4.0.0
markdown2==2.5.4
markdownify==1.1.0
marker-pdf==1.9.3
MarkupSafe==3.0.2
mcp==1.13.1
mdurl==0.1.2
mem0ai==0.1.93
metapub==0.6.4
modelscope==1.29.2
monotonic==1.6
MouseInfo==0.1.3
mpmath==1.3.0
multitasking==0.0.12
networkx==3.5
ninja==1.13.0
nodeenv==1.9.1
numpy==1.26.4
oauthlib==3.3.1
ollama==0.5.3
openai==1.107.1
opencv-contrib-python==4.6.0.66
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.11.3
packaging==24.2
paddlex==3.2.1
paddlepaddle==2.6.2
paddleocr==2.7.3 
pandas==2.3.2
patchright==1.55.1
pdfminer.six==20250506
pdfplumber==0.11.7
pdftext==0.6.3
pillow==10.4.0
platformdirs==4.4.0
playwright==1.55.0
portalocker==2.10.1
posthog==3.25.0
pre_commit==4.3.0
prettytable==3.16.0
proto-plus==1.26.1
protobuf==6.32.0
psutil==7.0.0
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pyclipper==1.3.0.post6
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
pydyf==0.11.0
pyee==13.0.0
PyGetWindow==0.0.9
Pygments==2.19.2
pymed==0.8.9
PyMsgBox==2.0.1
pyobjc-core==11.1
pyobjc-framework-Cocoa==11.1
pyobjc-framework-Quartz==11.1
pyparsing==3.2.3
pypdf==6.0.0
pypdfium2==4.30.0
pyperclip==1.9.0
pyphen==0.17.2
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-Levenshtein==0.27.1
python-magic==0.4.27
python-multipart==0.0.20
python-pptx==1.0.2
pytweening==1.2.0
pytz==2024.2
PyYAML==6.0.2
qdrant-client==1.15.1
RapidFuzz==3.14.1
referencing==0.36.2
regex==2024.4.28
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
rubicon-objc==0.5.2
ruff==0.13.0
s3transfer==0.14.0
safehttpx==0.1.6
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.1
screeninfo==0.8.1
semantic-version==2.10.0
sentence-transformers==5.1.0
sgmllib3k==1.0.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
sse-starlette==3.0.2
starlette==0.47.3
surya-ocr==0.16.7
sympy==1.14.0
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.6.0
tifffile==2025.9.9
tiktoken==0.11.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.22.0
tomlkit==0.13.3
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.56.1
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
Unidecode==1.4.0
uritemplate==4.2.0
urllib3==2.5.0
uuid7==0.1.0
uvicorn==0.35.0
virtualenv==20.34.0
waybackpy==3.0.6
wcwidth==0.2.13
weasyprint==63.1
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
wikipedia==1.4.0
xlsxwriter==3.2.5
yfinance==0.2.20
youtube-transcript-api==1.2.2
yt-dlp
zopfli==0.2.3.post1
zstandard==0.23.0
rank_bm25==0.2.2
accelerate
browser-use